+++ changelog.md
@@ -13,6 +13,7 @@
 
 - Optimized styling of list of directory services
 - Optimized handling of import through directory services
+- Moved import tasks from general Files in own object Import which is now also a directory listing object
 - Multiple code optimizations
 - Extended support for YouTube channel imports
 - More hooks